requirements.txt
streamlit
google-generativeai
python-docx
pdf2image
pillow